requirements.txt
django>=4.2.0
python-dotenv==1.0.0
openai>=1.0.0
twilio>=8.10.0
python-decouple==3.8